requirements.txt
pydantic[dotenv]
solara
flask_sock
fastapi
python-dotenv